refactor: plugin installation
