+++ console/packages/shared/src/index.ts
@@ -7,3 +7,4 @@ export * from "./states/editor";
 export * from "./states/plugin-tab";
 export * from "./states/comment-subject-ref";
 export * from "./states/backup";
+export * from "./states/plugin-installation-tabs";
