+++ console/src/modules/system/plugins/constants/index.ts
@@ -1,0 +1,2 @@
+export const PLUGIN_ALREADY_EXISTS_TYPE =
+  "https://halo.run/probs/plugin-alreay-exists";
