+++ console/src/modules/system/plugins/types/index.ts
@@ -1,0 +1,10 @@
+export interface PluginInstallationErrorResponse {
+  detail: string;
+  instance: string;
+  pluginName: string;
+  requestId: string;
+  status: number;
+  timestamp: string;
+  title: string;
+  type: string;
+}
